refactor: Event 관련 코드를 model 디렉토리로 이동
