+++ module-api/src/main/java/com/trybe/moduleapi/challenge/event/listener/ChallengePreferenceEventListener.java
@@ -1,6 +1,6 @@
 package com.trybe.moduleapi.challenge.event.listener;
 
-import com.trybe.moduleapi.challenge.event.ChallengeEvent;
+import com.trybe.moduleapi.challenge.event.model.ChallengeEvent;
 import com.trybe.moduleapi.challenge.event.ChallengeEventType;
 import com.trybe.modulecore.challenge.entity.Challenge;
 import com.trybe.modulecore.challenge.repository.preference.ChallengePreferenceCache;
