+++ module-api/src/main/java/com/trybe/moduleapi/challenge/event/model/ChallengeEvent.java
@@ -1,5 +1,6 @@
-package com.trybe.moduleapi.challenge.event;
+package com.trybe.moduleapi.challenge.event.model;
 
+import com.trybe.moduleapi.challenge.event.ChallengeEventType;
 import com.trybe.modulecore.challenge.entity.Challenge;
 import lombok.Getter;
