+++ module-api/src/main/java/com/trybe/moduleapi/challenge/event/pub/ChallengeEventPublisher.java
@@ -1,6 +1,6 @@
 package com.trybe.moduleapi.challenge.event.pub;
 
-import com.trybe.moduleapi.challenge.event.ChallengeEvent;
+import com.trybe.moduleapi.challenge.event.model.ChallengeEvent;
 
 public interface ChallengeEventPublisher {
     void publish(ChallengeEvent event);
